src/server/device_service/edge_downlink/requirements.txt
azure-iot-hub
Flask==3.1.0
